Scraping/requirements.txt
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
cssselect==1.2.0
greenlet==3.1.1
idna==3.10
jmespath==1.0.1
lxml==5.3.0
numpy==2.2.0
packaging==24.2
pandas==2.2.3
parsel==1.9.1
psycopg2==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.17.0
soupsieve==2.6
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
w3lib==2.2.1
